JavaScript code changes and improvements (#7)
* Update back to top button

JQuery is no longer used to further optimise the website. Back to top button functionality is now implemented only with HTML and CSS + a bit of JS to make it appear or disappear upon scrolling.

* Add animations.css file

Changes to back_to_top JavaScript file. Back to top button now has entrance and exit animation.

* Update search.php

Fix time displaying: use a 24-hour format instead of a 12-hour one.

* Update .gitignore

Ignore the Visual Studio Code files.

* Combine copy to clipboard and back to top files.

To lower the number of HTTP requests, copy to clipboard and back to top files are now combined.

* Update back_to_top.js

Using the new copy to clipboard function since the execCommand function is deprecated.

* Delete dropdown.js

It is no longer used. It was used for FAQ dropdown questions.

* Delete copy to clipboard script echo from all files

Since copy to clipboard was combined with back to top, this is no longer needed.

* Update back_to_top.js

Using the const keyword instead of let for HTML DOM elements.

* Update navigation files

Dialog element is now used instead of div.

Added menu dialog for responsiveness: the middle nav bar disappears and the menu icon appears next to the main logo.

* Update navigation.js

Use the function keyword instead of const for functions.

Change some variable names.

* Update back_to_top.js

Fix FAQ questions not working.

* Attempt to rewrite News page

Tried to pull data from the DB and dynamically change the content of the News page. It works, but I am not satisfied with it.